fix(quantization): backport refactor and MoE import-order fix
Backport upstream PR #14030 to releases/v0.26.0rc to repair cached
upstream MoE factory bindings while reorganizing quantization modules.

Retain v0.26 runner APIs, lazy configuration exports, and W4A8 group
metadata. Add regressions for stale factory bindings and release-specific
quantization metadata without introducing main-only prerequisites.

Backport of commit 73acadef668c4346558ae7061236742256cc8331.
Refs: #14030
Fixes: #15815
