+++ tests/ut/attention/a2/test_sfa_v1_precision.py
@@ -19,7 +19,7 @@
 import math
 import os
 import sys
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, patch
 
 import pytest
 import torch
@@ -467,6 +467,7 @@ def _run_precision_check(
     )
 
 
+@patch("vllm_ascend.quantization.utils.detect_quantization_method", new=MagicMock(return_value=None))
 @pytest.mark.parametrize("dtype", [torch.bfloat16, torch.float16])
 @pytest.mark.parametrize("model", ["deepseek-ai/DeepSeek-V3.2-Exp"])
 @pytest.mark.parametrize("batch_spec_name", list(BATCH_SPECS.keys()))
