+++ tests/ut/quantization/methods/test_w4a4_mxfp4.py
@@ -5,15 +5,15 @@ import torch.nn as nn
 
 from tests.ut.base import TestBase
 from tests.ut.quantization.conftest_quantization import create_mock_ascend_config, create_mock_vllm_config
-from vllm_ascend.quantization.methods.w4a4_mxfp4 import (
+from vllm_ascend.quantization.methods.w4a4.w4a4_mxfp4 import (
     AscendW4A4MXFP4DynamicFusedMoEMethod,
     AscendW4A4MXFP4DynamicLinearMethod,
 )
 
 
 class TestAscendW4A4MXFP4LinearMethod(TestBase):
-    @patch("vllm_ascend.quantization.methods.w4a4_mxfp4.ensure_mxfp4_linear_available")
-    @patch("vllm_ascend.quantization.methods.w4a4_mxfp4.get_current_vllm_config")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_mxfp4.ensure_mxfp4_linear_available")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_mxfp4.get_current_vllm_config")
     def setUp(self, mock_vllm, mock_ensure):
         mock_vllm.return_value = create_mock_vllm_config()
         mock_ensure.return_value = None
@@ -41,7 +41,7 @@ class TestAscendW4A4MXFP4LinearMethod(TestBase):
         self.assertEqual(layer.weight.shape, (128, 128))
         self.assertEqual(layer.weight_scale.shape[0], 4)
 
-    @patch("vllm_ascend.quantization.methods.w4a4_mxfp4.torch_npu")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_mxfp4.torch_npu")
     def test_apply_3d_input(self, mock_npu):
         mock_npu.npu_dynamic_mx_quant.return_value = (
             torch.randint(0, 255, (32, 128), dtype=torch.uint8),
@@ -62,9 +62,9 @@ class TestAscendW4A4MXFP4MoEMethod(TestBase):
     hidden_size = 128
     intermediate_size = 256
 
-    @patch("vllm_ascend.quantization.methods.w4a4_mxfp4.ensure_mxfp4_moe_available")
-    @patch("vllm_ascend.quantization.methods.w4a4_mxfp4.get_current_vllm_config")
-    @patch("vllm_ascend.quantization.methods.w4a4_mxfp4.get_ascend_config")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_mxfp4.ensure_mxfp4_moe_available")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_mxfp4.get_current_vllm_config")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_mxfp4.get_ascend_config")
     def setUp(self, mock_ascend, mock_vllm, mock_ensure):
         mock_vllm.return_value = create_mock_vllm_config()
         mock_ascend.return_value = create_mock_ascend_config()
@@ -103,9 +103,9 @@ class TestAscendW4A4MXFP4MoEMethod(TestBase):
         self.assertEqual(layer.w13_weight.shape, (8, 64, 256))
         self.assertEqual(layer.w13_weight_scale.shape, (8, 2, 256, 2))
 
-    @patch("vllm_ascend.quantization.methods.w4a4_mxfp4.torch_npu")
-    @patch("vllm_ascend.quantization.methods.w4a4_mxfp4._EXTRA_CTX")
-    @patch("vllm_ascend.quantization.methods.w4a4_mxfp4.select_experts")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_mxfp4.torch_npu")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_mxfp4._EXTRA_CTX")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_mxfp4.select_experts")
     def test_apply_full_params(self, mock_select, mock_ctx, mock_npu):
         tokens = 4
         layer = nn.Module()
