+++ tests/ut/quantization/methods/test_w8a8_pdmix.py
@@ -5,19 +5,15 @@ import torch
 from tests.ut.base import TestBase
 from tests.ut.quantization.conftest_quantization import create_mock_vllm_config
 from vllm_ascend.quantization.methods import (
-    AscendW8A8LinearMethod,
     AscendW8A8PDMixFusedMoeMethod,
     AscendW8A8PDMixLinearMethod,
 )
 
 
 class TestAscendW8A8PDMixLinearScheme(TestBase):
-    def setUp(self):
-        self.method = AscendW8A8LinearMethod()
-
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.AscendW8A8LinearMethod")
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.AscendW8A8DynamicLinearMethod")
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.get_current_vllm_config")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.AscendW8A8LinearMethod")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.AscendW8A8DynamicLinearMethod")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.get_current_vllm_config")
     def test_get_weight_delegates_to_static(self, mock_vllm_config, mock_dynamic_cls, mock_static_cls):
         mock_vllm_config.return_value = create_mock_vllm_config(kv_transfer_config=None)
         mock_dynamic_instance = MagicMock()
@@ -31,9 +27,9 @@ class TestAscendW8A8PDMixLinearScheme(TestBase):
             mock_static_instance.get_weight.assert_called_with(input_size, output_size, torch.bfloat16)
             mock_dynamic_instance.get_weight.assert_not_called()
 
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.AscendW8A8LinearMethod")
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.AscendW8A8DynamicLinearMethod")
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.get_current_vllm_config")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.AscendW8A8LinearMethod")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.AscendW8A8DynamicLinearMethod")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.get_current_vllm_config")
     def test_get_pertensor_param_delegates_to_static(self, mock_vllm_config, mock_dynamic_cls, mock_static_cls):
         mock_vllm_config.return_value = create_mock_vllm_config(kv_transfer_config=None)
         mock_dynamic_instance = MagicMock()
@@ -46,9 +42,9 @@ class TestAscendW8A8PDMixLinearScheme(TestBase):
         mock_static_instance.get_pertensor_param.assert_called_once_with(torch.bfloat16)
         mock_dynamic_instance.get_pertensor_param.assert_not_called()
 
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.AscendW8A8LinearMethod")
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.AscendW8A8DynamicLinearMethod")
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.get_current_vllm_config")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.AscendW8A8LinearMethod")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.AscendW8A8DynamicLinearMethod")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.get_current_vllm_config")
     def test_get_perchannel_param_delegates_to_static(self, mock_vllm_config, mock_dynamic_cls, mock_static_cls):
         mock_vllm_config.return_value = create_mock_vllm_config(kv_transfer_config=None)
         mock_dynamic_instance = MagicMock()
@@ -61,9 +57,9 @@ class TestAscendW8A8PDMixLinearScheme(TestBase):
         mock_static_instance.get_perchannel_param.assert_called_once_with(128, torch.bfloat16)
         mock_dynamic_instance.get_perchannel_param.assert_not_called()
 
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.AscendW8A8LinearMethod")
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.AscendW8A8DynamicLinearMethod")
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.get_current_vllm_config")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.AscendW8A8LinearMethod")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.AscendW8A8DynamicLinearMethod")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.get_current_vllm_config")
     def test_apply_uses_static_for_kv_consumer(self, mock_vllm_config, mock_dynamic_cls, mock_static_cls):
         mock_vllm_config.return_value = create_mock_vllm_config(kv_transfer_config=None)
         mock_static_instance = MagicMock()
@@ -79,9 +75,9 @@ class TestAscendW8A8PDMixLinearScheme(TestBase):
         mock_static_instance.apply.assert_called_once()
         mock_dynamic_instance.apply.assert_not_called()
 
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.AscendW8A8LinearMethod")
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.AscendW8A8DynamicLinearMethod")
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.get_current_vllm_config")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.AscendW8A8LinearMethod")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.AscendW8A8DynamicLinearMethod")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.get_current_vllm_config")
     def test_apply_uses_dynamic_for_non_kv_consumer(self, mock_vllm_config, mock_dynamic_cls, mock_static_cls):
         mock_vllm_config.return_value = create_mock_vllm_config(kv_transfer_config=None)
         mock_dynamic_instance = MagicMock()
@@ -97,9 +93,9 @@ class TestAscendW8A8PDMixLinearScheme(TestBase):
         mock_dynamic_instance.apply.assert_called_once()
         mock_static_instance.apply.assert_not_called()
 
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.AscendW8A8LinearMethod")
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.AscendW8A8DynamicLinearMethod")
-    @patch("vllm_ascend.quantization.methods.w8a8_pdmix.get_current_vllm_config")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.AscendW8A8LinearMethod")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.AscendW8A8DynamicLinearMethod")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_pdmix.get_current_vllm_config")
     def test_process_weights_after_loading_sets_is_kv_consumer(
         self, mock_vllm_config, mock_dynamic_cls, mock_static_cls
     ):
@@ -118,9 +114,9 @@ class TestAscendW8A8PDMixLinearScheme(TestBase):
 
 
 class TestAscendW8A8PDMixMoEScheme(TestBase):
-    @patch("vllm_ascend.quantization.methods.w8a8_dynamic.get_mc2_group")
-    @patch("vllm_ascend.quantization.methods.w8a8_dynamic.get_current_vllm_config")
-    @patch("vllm_ascend.quantization.methods.w8a8_dynamic.get_ascend_config")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_dynamic.get_mc2_group")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_dynamic.get_current_vllm_config")
+    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_dynamic.get_ascend_config")
     def test_get_dynamic_quant_param(self, mock_ascend, mock_vllm, mock_mc2):
         mock_mc2.side_effect = AttributeError()
         mock_vllm.return_value = create_mock_vllm_config()
