+++ vllm_ascend/quantization/__init__.py
@@ -24,24 +24,24 @@ paths and circular imports during startup.
 from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
-    from .compressed_tensors_config import AscendCompressedTensorsConfig
-    from .fp8_config import AscendFp8Config
-    from .modelslim_config import AscendModelSlimConfig
+    from .configs.compressed_tensors_config import AscendCompressedTensorsConfig
+    from .configs.fp8_config import AscendFp8Config
+    from .configs.modelslim_config import AscendModelSlimConfig
 
 __all__ = ["AscendModelSlimConfig", "AscendCompressedTensorsConfig", "AscendFp8Config"]
 
 
 def __getattr__(name: str) -> Any:
     if name == "AscendModelSlimConfig":
-        from .modelslim_config import AscendModelSlimConfig
+        from .configs.modelslim_config import AscendModelSlimConfig
 
         return AscendModelSlimConfig
     if name == "AscendCompressedTensorsConfig":
-        from .compressed_tensors_config import AscendCompressedTensorsConfig
+        from .configs.compressed_tensors_config import AscendCompressedTensorsConfig
 
         return AscendCompressedTensorsConfig
     if name == "AscendFp8Config":
-        from .fp8_config import AscendFp8Config
+        from .configs.fp8_config import AscendFp8Config
 
         return AscendFp8Config
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
