+++ vllm_ascend/quantization/configs/__init__.py
@@ -1,0 +1,27 @@
+#
+# Copyright (c) 2026 Huawei Technologies Co., Ltd. All Rights Reserved.
+# This file is a part of the vllm-ascend project.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+"""Quantization configuration classes for Ascend NPU."""
+
+from .compressed_tensors_config import AscendCompressedTensorsConfig
+from .fp8_config import AscendFp8Config
+from .modelslim_config import AscendModelSlimConfig
+
+__all__ = [
+    "AscendCompressedTensorsConfig",
+    "AscendFp8Config",
+    "AscendModelSlimConfig",
+]
