+++ vllm_ascend/quantization/methods/w4a8/w4a8.py
@@ -32,8 +32,8 @@ from vllm_ascend.ops.fused_moe.experts_selector import select_experts
 from vllm_ascend.ops.fused_moe.moe_runtime_args import build_fused_experts_input
 from vllm_ascend.utils import COMPRESSED_TENSORS_METHOD, maybe_trans_nz
 
-from .base import AscendLinearScheme, AscendMoEScheme, QuantType, get_moe_num_logical_experts
-from .registry import register_scheme
+from ..base import AscendLinearScheme, AscendMoEScheme, QuantType, get_moe_num_logical_experts
+from ..registry import register_scheme
 
 
 @register_scheme("W4A8_DYNAMIC", "linear")
@@ -239,7 +239,7 @@ class AscendW4A8DynamicFusedMoEMethod(AscendMoEScheme):
 
     This method supports four MoE weight formats: three generated by
     msModelSlim and one generated by LLM-Compressor. The LLM-Compressor path
-    is selected when ``ascend_quant_method`` in ``quant_description`` is
+    is selected when the quantization config's ``get_name()`` returns
     ``COMPRESSED_TENSORS_METHOD``. Otherwise, the msModelSlim path is used.
     msModelSlim layouts are first distinguished by
     ``quant_description["version"] == "1.0.0"``; for version ``"1.0.0"``,
@@ -365,7 +365,7 @@ class AscendW4A8DynamicFusedMoEMethod(AscendMoEScheme):
         # NOTE: new quantize weights: 2 int4 pack into int8
         self.new_quant_version = quant_version == "1.0.0"
 
-        self.quant_method = vllm_config.quant_config.quant_description.get("ascend_quant_method", "")
+        self.quant_method = vllm_config.quant_config.get_name()
         if self.quant_method == COMPRESSED_TENSORS_METHOD:
             self.weight_strategy = vllm_config.quant_config.quant_description.get("weight_strategy", "group")
